subsamplers: refactor file into directory, file might get very long (#82)
* subsamplers: refactor file into directory, file might get very long

* fix lint
